requirements.txt
requests==2.31.0
python-dotenv==1.0.0
shopify-python-api>=12.4.0
typing-extensions>=4.9.0
schedule==1.2.1
pandas==2.2.0
openpyxl==3.1.2 